Fixed all compiler warnings

+++ src/gui/widgets/Carousel.h
@@ -18,35 +18,43 @@ namespace dmt {
 namespace gui {
 namespace panels {
 //==============================================================================
-class Carousel : public juce::Component {
+class Carousel : public juce::Component
+{
 public:
-  Carousel() : index(0) {}
-  void next() {
+  Carousel()
+    : index(0)
+  {
+  }
+  void next()
+  {
     panels[index]->setVisible(false);
     index++;
     if (index >= panels.size())
-      index -= panels.size();
+      index -= (int)panels.size();
     panels[index]->setVisible(true);
     repaint();
   }
-  void previous() {
+  void previous()
+  {
     panels[index]->setVisible(false);
     index--;
     if (index < 0)
-      index += panels.size();
+      index += (int)panels.size();
     panels[index]->setVisible(true);
     repaint();
   }
-  void init() {
-    for (auto &panel : panels) {
+  void init()
+  {
+    for (auto& panel : panels) {
       panel->setCallbacks([this]() { next(); }, [this]() { previous(); });
       addChildComponent(*panel);
     }
     panels[index]->setVisible(true);
   }
 
-  void resized() override {
-    for (auto &panel : panels) {
+  void resized() override
+  {
+    for (auto& panel : panels) {
       panel->setBoundsRelative(0.0f, 0.0f, 1.0f, 1.0f);
     }
   }
